Working Part 1
Fuck tonne of changes

+++ Arduino/libraries/Task3Control1/Task3Control1.h
@@ -51,7 +51,7 @@ class control{
         }
     
     
-        int getInterCheck()             //returns set check interval
+        unsigned long getInterCheck()             //returns set check interval
         {
             return check_interval;
         }
